requirements.txt
azure-cognitiveservices-speech
dspy-ai
openai
pyairtable
python-dotenv
requests
